Add files via upload
Globally, people are living longer. Today most people can expect to live into their sixties and beyond. There is a growing proportion of older people in every country in the world. One in six people in the world will be over 60 by 2030.

The world's elderly population will double (2.1 billion) by 2050. Low- and middle-income countries are experiencing the greatest change. By 2050, four in six people in the world will be over 60 years in low- and middle-income countries.

Longer life brings with it opportunities, not only for older people and their families but also for society. 
When people can spend these extra years in good health and in an environment that supports them, they can do more of what they value than they could when they were younger. But, if these added years are dominated by declines in physical and mental capacity, the implications for older people and society are more negative.